moved utils to lib folder
