+++ src/lib/mappingUtils.js
